+++ core/validatorapi/router_internal_test.go
@@ -20,6 +20,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"math/rand"
 	"net/http"
 	"net/http/httptest"
 	"net/http/httputil"
@@ -27,6 +28,7 @@ import (
 	"strings"
 	"testing"
 
+	eth2client "github.com/attestantio/go-eth2-client"
 	eth2api "github.com/attestantio/go-eth2-client/api"
 	eth2v1 "github.com/attestantio/go-eth2-client/api/v1"
 	eth2http "github.com/attestantio/go-eth2-client/http"
@@ -39,7 +41,9 @@ import (
 
 	"github.com/obolnetwork/charon/app/errors"
 	"github.com/obolnetwork/charon/app/eth2wrap"
+	"github.com/obolnetwork/charon/eth2util/eth2exp"
 	"github.com/obolnetwork/charon/testutil"
+	"github.com/obolnetwork/charon/testutil/beaconmock"
 )
 
 const (
@@ -543,6 +547,80 @@ func TestRouter(t *testing.T) {
 
 		testRouter(t, handler, callback)
 	})
+}
+
+func TestBeaconCommitteeSubscriptionsV2(t *testing.T) {
+	ctx := context.Background()
+
+	const (
+		slotA = 123
+		slotB = 456
+		vIdxA = 1
+		vIdxB = 2
+		vIdxC = 3
+	)
+
+	aggregators := map[eth2p0.Slot]eth2p0.ValidatorIndex{
+		slotA: vIdxA,
+		slotB: vIdxB,
+	}
+
+	bmock, err := beaconmock.New(beaconmock.WithAttestationAggregation(aggregators))
+	require.NoError(t, err)
+
+	expected := []*eth2exp.BeaconCommitteeSubscriptionResponse{
+		{ValidatorIndex: vIdxA, IsAggregator: true},
+		{ValidatorIndex: vIdxB, IsAggregator: true},
+		{ValidatorIndex: vIdxC, IsAggregator: false},
+	}
+
+	subs := []*eth2exp.BeaconCommitteeSubscription{
+		{
+			Slot:             slotA,
+			ValidatorIndex:   vIdxA,
+			CommitteesAtSlot: rand.Uint64(),
+			CommitteeIndex:   eth2p0.CommitteeIndex(rand.Uint64()),
+			SlotSignature:    testutil.RandomEth2Signature(),
+		},
+		{
+			Slot:             slotB,
+			ValidatorIndex:   vIdxB,
+			CommitteesAtSlot: rand.Uint64(),
+			CommitteeIndex:   eth2p0.CommitteeIndex(rand.Uint64()),
+			SlotSignature:    testutil.RandomEth2Signature(),
+		},
+		{
+			Slot:             slotA,
+			ValidatorIndex:   vIdxC,
+			CommitteesAtSlot: rand.Uint64(),
+			CommitteeIndex:   eth2p0.CommitteeIndex(rand.Uint64()),
+			SlotSignature:    testutil.RandomEth2Signature(),
+		},
+	}
+
+	vapi, err := NewComponentInsecure(t, bmock, 0)
+	require.NoError(t, err)
+
+	vapiRouter, err := NewRouter(vapi, bmock)
+	require.NoError(t, err)
+
+	server := httptest.NewServer(vapiRouter)
+	defer server.Close()
+
+	var eth2Svc eth2client.Service
+	eth2Svc, err = eth2http.New(ctx,
+		eth2http.WithLogLevel(1),
+		eth2http.WithAddress(server.URL),
+	)
+	require.NoError(t, err)
+
+	eth2Cl, err := eth2wrap.NewWrapHTTP(eth2Svc)
+	require.NoError(t, err)
+
+	// Submit subscriptions to beaconmock through validatorapi.
+	actual, err := eth2Cl.SubmitBeaconCommitteeSubscriptionsV2(ctx, subs)
+	require.NoError(t, err)
+	require.Equal(t, expected, actual)
 }
 
 // testRouter is a helper function to test router endpoints with an eth2http client. The outer test
