+++ testutil/beaconmock/beaconmock_test.go
@@ -17,20 +17,13 @@ package beaconmock_test
 
 import (
 	"context"
-	"math/rand"
-	"net/http/httptest"
 	"testing"
 	"time"
 
-	eth2client "github.com/attestantio/go-eth2-client"
-	eth2http "github.com/attestantio/go-eth2-client/http"
 	eth2p0 "github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/jonboulle/clockwork"
 	"github.com/stretchr/testify/require"
 
-	"github.com/obolnetwork/charon/app/eth2wrap"
-	"github.com/obolnetwork/charon/core/validatorapi"
-	"github.com/obolnetwork/charon/eth2util/eth2exp"
 	"github.com/obolnetwork/charon/testutil"
 	"github.com/obolnetwork/charon/testutil/beaconmock"
 )
@@ -70,78 +63,4 @@ func TestAttestationData(t *testing.T) {
 	attData, err := bmock.AttestationData(context.Background(), 1, 2)
 	require.NoError(t, err)
 	testutil.RequireGoldenJSON(t, attData)
-}
-
-func TestBeaconCommitteeSubscriptions(t *testing.T) {
-	ctx := context.Background()
-
-	const (
-		slotA = 123
-		slotB = 456
-		vIdxA = 1
-		vIdxB = 2
-		vIdxC = 3
-	)
-
-	aggregators := map[eth2p0.Slot]eth2p0.ValidatorIndex{
-		slotA: vIdxA,
-		slotB: vIdxB,
-	}
-
-	bmock, err := beaconmock.New(beaconmock.WithAttestationAggregation(aggregators))
-	require.NoError(t, err)
-
-	expected := []*eth2exp.BeaconCommitteeSubscriptionResponse{
-		{ValidatorIndex: vIdxA, IsAggregator: true},
-		{ValidatorIndex: vIdxB, IsAggregator: true},
-		{ValidatorIndex: vIdxC, IsAggregator: false},
-	}
-
-	subs := []*eth2exp.BeaconCommitteeSubscription{
-		{
-			Slot:             slotA,
-			ValidatorIndex:   vIdxA,
-			CommitteesAtSlot: rand.Uint64(),
-			CommitteeIndex:   eth2p0.CommitteeIndex(rand.Uint64()),
-			SlotSignature:    testutil.RandomEth2Signature(),
-		},
-		{
-			Slot:             slotB,
-			ValidatorIndex:   vIdxB,
-			CommitteesAtSlot: rand.Uint64(),
-			CommitteeIndex:   eth2p0.CommitteeIndex(rand.Uint64()),
-			SlotSignature:    testutil.RandomEth2Signature(),
-		},
-		{
-			Slot:             slotA,
-			ValidatorIndex:   vIdxC,
-			CommitteesAtSlot: rand.Uint64(),
-			CommitteeIndex:   eth2p0.CommitteeIndex(rand.Uint64()),
-			SlotSignature:    testutil.RandomEth2Signature(),
-		},
-	}
-
-	vapi, err := validatorapi.NewComponentInsecure(t, bmock, 0)
-	require.NoError(t, err)
-
-	vapiRouter, err := validatorapi.NewRouter(vapi, bmock)
-	require.NoError(t, err)
-
-	server := httptest.NewServer(vapiRouter)
-	defer server.Close()
-
-	var eth2Svc eth2client.Service
-	eth2Svc, err = eth2http.New(ctx,
-		eth2http.WithLogLevel(1),
-		eth2http.WithAddress(server.URL),
-	)
-	require.NoError(t, err)
-
-	eth2Cl, err := eth2wrap.NewWrapHTTP(eth2Svc)
-	require.NoError(t, err)
-
-	// Submit subscriptions to beaconmock through validatorapi.
-	actual, err := eth2Cl.SubmitBeaconCommitteeSubscriptionsV2(ctx, subs)
-	require.NoError(t, err)
-	require.Equal(t, expected, actual)
 }
